Bundle the auth lifecycle into a single named-operation unit
Replaces the scattered auth-flip code (clearUser in clear-user.ts,
restorePersistedUserData in local-cache.ts, the profile preload/recovery
inlined in the _user loader) with a single class in src/lib/auth-lifecycle.ts
that exposes three async operations, each self-documenting via its name:

  - registerPersistedMirrorsAtBootstrap() — called from main.tsx
  - clearAllUserDataOnIdentityChange()   — called from auth-context on
                                            any identity change
  - loadAllTheRequiredUserDataAfterNewLogin(userId) — called from the
                                            _user route loader

Runtime behaviour is unchanged (still fire-and-forget on the clear, still
the same 3-step recovery in the loader). The wins are structural:

  - The contract between auth events / route loaders / collection
    cleanup lives in one file with explicit named methods, so changing
    one timing assumption requires updating its named caller too.
  - Two new failed() checks at the load-after-login site fire in
    scenetest runs when the contract is violated: hitting the profile
    recovery branch means clearAllUserDataOnIdentityChange didn't run in
    time, and falling through to the missing-profile throw with seed
    data means the handle_new_user trigger / backfill didn't run. The
    production fallbacks (recovery + throw) still execute so users get
    the same behaviour as before; the failed() calls are stripped from
    the production bundle.

local-cache.ts retains its mirror configuration (PERSISTED_COLLECTIONS,
attachMirror, hydrateFromStorage, clearPersistedUserData,
hasSupabaseSessionToken) — auth-lifecycle composes them.
clear-user.ts is deleted; its contents moved into the class.
